Added Media Layer Editor.
Changed to dotnet 7 for blazor stuff and blazorwebview.

Added some path constants and moved some assets to wwwroot/assets so it's possible for blazor to have access to some of assets.

Fixed possible small bugs.

Rewrote the download media to be able to use multiple threads id user have more threads available from screenscraper.fr.

Changed so that game media gets added with game id and not deleted until program close.
There's a settings for not deleting temp folder on exit but it's not in the UI right know.
It will also return media assets if they already exists in temp folder without doing a request to screenscraper.fr.

It should also now correctly choose the game media with the correct region order.